# packages/contracts/openapi/prism-v1.yaml
openapi: 3.1.0
info:
  title: PRism API
  version: 1.0.0
  description: REST API served by services/gateway
servers:
  - url: http://localhost:3001
paths:
...
  /api/security/findings:
    get:
      summary: List security findings (filtered, paginated)
      parameters:
        - name: severity
          in: query
          schema: { type: string }
        - name: repo
          in: query
          schema: { type: string }
        - name: q
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, minimum: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 100 }
    post:
      summary: Create security finding
...
  /api/performance/findings:
    get:
      summary: List performance findings (filtered, paginated)
      parameters:
        - name: severity
          in: query
          schema: { type: string }
        - name: type
          in: query
          schema: { type: string }
        - name: repo
          in: query
          schema: { type: string }
        - name: q
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, minimum: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 100 }
...